Removed indirection of random and latest invokers. Changed directory structure.
